LookMaker updates

+++ major/src/main/java/com/app/shop/repository/ClothSetRepository.java
@@ -3,6 +3,6 @@ package com.app.shop.repository;
 import com.app.shop.model.ClothSet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClothSetRepository extends JpaRepository<ClothSet, Long> {
+public interface ClothSetRepository extends JpaRepository<ClothSet, Integer> {
     ClothSet findClothSetById(Integer id);
 }
